+++ GLAndroid/src/com/gooble/logic/app/entity/Puzzle.java
@@ -3,7 +3,7 @@ package com.gooble.logic.app.entity;
 public class Puzzle extends Entity{
 
    private String name;
-   private Integer mainvariableid;
+   private Long mainvariableid;
    
    public String getName() {
       return name;
@@ -11,10 +11,10 @@ public class Puzzle extends Entity{
    public void setName(String name) {
       this.name = name;
    }
-   public Integer getMainvariableid() {
+   public Long getMainvariableid() {
       return mainvariableid;
    }
-   public void setMainvariableid(Integer mainvariableid) {
+   public void setMainvariableid(Long mainvariableid) {
       this.mainvariableid = mainvariableid;
    }
